fix: increase size of offsets

+++ include/butter/types.h
@@ -156,10 +156,10 @@ typedef struct {
   butter_pipeline_t pipeline;
   vk_buffer_t vertex_buffer;
   u32 vertex_count;
-  u32 vertex_offset;
+  u64 vertex_offset;
   vk_buffer_t index_buffer;
   u32 index_count;
-  u32 index_offset;
+  u64 index_offset;
   vk_index_type_t index_type;
   const butter_descriptor_set_t *descriptor_sets;
   u32 descriptor_set_count;
